test(backend): retire *_coverage.py and merge into primary test files (Issue #89)
